make the script a typescript one and centralize the references
