+++ package.json
@@ -94,6 +94,7 @@
     "sass": "^1.89.0",
     "supertest": "^7.0.0",
     "testcontainers": "^10.23.0",
+    "ts-node": "^10.9.2",
     "typescript": "^5"
   }
 }
